requirements.txt
django==4.1.7
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dotenv==0.21.1
django-mptt==0.14.0
django-bootstrap-v5==1.0.11
jupyterlab==3.6.1
notebook==6.5.3
django-extensions==3.2.1
pillow==9.5.0